ocrBackend/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.0.2
google-api-core==1.14.3
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
googleapis-common-protos==1.6.0
grpcio==1.24.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pymongo==3.9.0
pytz==2019.3
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.16.0
flask_restful==0.3.7
flask_imgur==0.1
flask_jwt_extended==3.24.0
opencv-python==4.1.2.30
baidu-aip==2.2.18.0
imutils==0.5.3


